perf: use decrement instead of splice/indexOf
In _onAll, individual elements were being removed from an array to track
if they had been run. In reality, all that's needed is to know when
*all* of them have been run. This can be acheived by simply getting
the original array length, and decrementing it upon completion of an
element, with a stop condition when the count is zero.

In a small benchmark, this increased speeds for resolution of a graph by
roughly 4.5x. This was clearly the lowest hanging fruit in terms of
performance of asyncgrpah.